Add Shell 2.0 overrides from TMP teams-stardust-ui.
