+++ src/components/States/States/Default.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { TeamsTheme } from "../../../lib/withTheme";
+import { TeamsTheme } from "../../../themes";
 import { IState, State } from "./State";
 
 const DEFAULT_STRINGS = {
