+++ src/components/Toolbar/ToolbarTheme.tsx
@@ -12,11 +12,9 @@ import {
   ComponentVariablesObject,
 } from "@fluentui/styles";
 
-import {
-  teamsNextVariableAssignments,
-  TeamsTheme,
-  themes,
-} from "../../lib/withTheme";
+import { teamsNextVariableAssignments, themes } from "../../lib/withTheme";
+
+import { TeamsTheme } from "../../themes";
 
 export interface IToolbarThemeProps {
   globalTheme: ThemePrepared;
