+++ src/stories/States.stories.tsx
@@ -1,14 +1,11 @@
 import React from "react";
-import { withKnobs, object } from "@storybook/addon-knobs";
-import { withA11y } from "@storybook/addon-a11y";
+import { object } from "@storybook/addon-knobs";
 
 import States, { StatesOptions } from "../components/States";
-import { StorybookThemeProvider } from "../lib/withTheme";
 
 export default {
   title: "Components/States",
   component: States,
-  decorators: [withKnobs, withA11y],
 };
 
 const statesKnobGroupID = "States";
@@ -19,9 +16,7 @@ const statesKnobGroupID = "States";
 const defaultConfig = { option: StatesOptions.Default };
 export const Default = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", defaultConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", defaultConfig, statesKnobGroupID)} />
   );
 };
 
@@ -31,9 +26,7 @@ export const Default = () => {
 const welcomeConfig = { option: StatesOptions.Welcome };
 export const Welcome = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", welcomeConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", welcomeConfig, statesKnobGroupID)} />
   );
 };
 
@@ -43,9 +36,7 @@ export const Welcome = () => {
 const helloConfig = { option: StatesOptions.Hello };
 export const Hello = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", helloConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", helloConfig, statesKnobGroupID)} />
   );
 };
 
@@ -55,9 +46,7 @@ export const Hello = () => {
 const emptyConfig = { option: StatesOptions.Empty };
 export const Empty = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", emptyConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", emptyConfig, statesKnobGroupID)} />
   );
 };
 
@@ -67,9 +56,7 @@ export const Empty = () => {
 const errorConfig = { option: StatesOptions.Error };
 export const Error = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", errorConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", errorConfig, statesKnobGroupID)} />
   );
 };
 
@@ -79,9 +66,7 @@ export const Error = () => {
 const thanksConfig = { option: StatesOptions.Thanks };
 export const Thanks = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", thanksConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", thanksConfig, statesKnobGroupID)} />
   );
 };
 
@@ -175,8 +160,6 @@ const customConfig = {
 };
 export const Custom = () => {
   return (
-    <StorybookThemeProvider>
-      <States {...object("Configuration", customConfig, statesKnobGroupID)} />
-    </StorybookThemeProvider>
+    <States {...object("Configuration", customConfig, statesKnobGroupID)} />
   );
 };
