+++ src/stories/Table.stories.tsx
@@ -1,14 +1,10 @@
 import React from "react";
-import { withKnobs, object, boolean } from "@storybook/addon-knobs";
+import { object, boolean } from "@storybook/addon-knobs";
 import Table from "../components/Table";
-import { withA11y } from "@storybook/addon-a11y";
-
-import { StorybookThemeProvider } from "../lib/withTheme";
 
 export default {
   title: "Components/Table",
   component: Table,
-  decorators: [withKnobs, withA11y],
 };
 
 const tableKnobGroupID = "Table";
@@ -80,12 +76,10 @@ export const KitchenSink = () => {
     },
   };
   return (
-    <StorybookThemeProvider>
-      <Table
-        truncate={boolean("Truncate", false, tableKnobGroupID)}
-        selectable={boolean("Selectable", true, tableKnobGroupID)}
-        {...object("Configuration", tableConfig, tableKnobGroupID)}
-      />
-    </StorybookThemeProvider>
+    <Table
+      truncate={boolean("Truncate", false, tableKnobGroupID)}
+      selectable={boolean("Selectable", true, tableKnobGroupID)}
+      {...object("Configuration", tableConfig, tableKnobGroupID)}
+    />
   );
 };
