+++ src/themes/constants.ts
@@ -1,0 +1,5 @@
+export enum TeamsTheme {
+  HighContrast = "teamsHighContrastTheme",
+  Dark = "teamsDarkTheme",
+  Default = "teamsTheme",
+}
