+++ src/themes/index.ts
@@ -1,0 +1,12 @@
+// Based upon TMP's `teams-stardust-ui`
+
+export { TeamsTheme } from "./constants";
+
+// Teams high contrast theme overrides
+export { highContrastThemeOverrides } from "./teams-high-contrast/theme";
+
+// Teams defaultV2 theme overrides
+export { defaultV2ThemeOverrides } from "./teams-v2/theme";
+
+// Teams darkV2 theme overrides
+export { darkV2ThemeOverrides } from "./teams-dark-v2/theme";
