+++ src/themes/teams-high-contrast/components/Card/card-styles.ts
@@ -1,0 +1,18 @@
+import {
+  ComponentSlotStylesPrepared,
+  TeamsThemeStylesProps,
+} from "@fluentui/react-northstar";
+
+export const cardStyles: ComponentSlotStylesPrepared<
+  NonNullable<Partial<TeamsThemeStylesProps["Card"]>>,
+  any
+> = {
+  root: ({
+    theme: {
+      siteVariables: { colorScheme, colors },
+    },
+    variables: {},
+  }) => {
+    return {};
+  },
+};
