+++ src/themes/teams-high-contrast/site-variables.ts
@@ -1,0 +1,12 @@
+import { TeamsTheme } from "../constants";
+
+export default {
+  theme: TeamsTheme.HighContrast,
+  colorScheme: {
+    elevations: {
+      4: "none",
+      8: "none",
+      16: "none",
+    },
+  },
+};
