+++ src/themes/teams-high-contrast/static-styles.ts
@@ -1,0 +1,11 @@
+/**
+ * Custom static styles.
+ *
+ * ! Avoid adding anything to this export, unless absolutely necessary.
+ * ! Rather, refactor these definitions into component-specific styles.
+ */
+export const staticStyles = [
+  `#LivePersonaCardRootElement .customScrollBar::-webkit-scrollbar-thumb {
+      background-color: #fff;
+    }`,
+];
